pyproject.toml
[project]
name = "data_guide"
version = "0.1.0"
description = "General-purpose EDA tool with pluggable connectors and multiple output formats"
authors = [{ name = "Zach" }]
requires-python = ">=3.11"
dependencies = [
  "pandas",
  "numpy",
  "scipy",
  "matplotlib",
  "seaborn",
  "missingno",
  "statsmodels",
  "scikit-learn",
  "typer",
  "pyyaml",
  "tabulate>=0.9.0",
  "markdown",
  "beautifulsoup4",
  "openpyxl",
  "sqlalchemy",
]

[project.optional-dependencies]
snowflake = ["snowflake-connector-python"]
phone = ["phonenumbers"]
geo = ["geopandas", "esda"]
pdf = ["pdfplumber"]
dev = ["pytest", "pytest-cov", "pre-commit", "ruff"]

[project.scripts]
guide = "cli:app"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
